+++ CHANGES.md
@@ -186,3 +186,6 @@
 - Rexton: fix Connexx checker for v9.13.0.1436 by changing target market
 - Audio Service: fix Connexx checker for v9.13.0.1420 by changing target market
 - Signia/Rexton/Audio Service/A&M: script tweak for easy changing target market and base version
+
+### Pre-release (work-in-progress)
+- Phonak/Unitron/Hansaton/Oticon/Bernafon/Sonic/Philips: script tweak for easy changing target market and base version